[config-nodemgr]: Modified to track config db
Deployments where config and analytics db processes are seprated, we
were only tracking tracking analytics db status. This patch takes care
of tracking config db status.
Separate bug needs to be opened for UI to read config cassandra status
UVEs and display the status accordingly.

Closes-Bug: #1705173
