+++ src/analytics/database/database.sandesh
@@ -16,7 +16,8 @@ include "base/sandesh/process_info.sandesh"
 struct DatabaseUsageStats {
     1: u64                                 disk_space_used_1k
     2: u64                                 disk_space_available_1k
-    3: u64                                 analytics_db_size_1k
+    3: optional u64                        analytics_db_size_1k
+    4: optional u64                        config_db_size_1k
 }
 
 /**
