+++ src/config/utils/contrail-status.py
@@ -393,6 +393,17 @@ def supervisor_status(nodetype, options):
     elif nodetype == 'config':
         print "== Contrail Config =="
         check_status('supervisor-config', options)
+        for svc_name in CONTRAIL_SERVICES[nodetype][init_sys_used]:
+            check_status(svc_name, options)
+        db = package_installed('contrail-openstack-database')
+        config_db = package_installed('contrail-database-common')
+        if not db and config_db:
+            print "== Contrail Config Database=="
+            initd_svc = init_sys_used != 'systemd'
+            check_return_code = True
+            check_svc('contrail-database', initd_svc=initd_svc,
+                check_return_code=check_return_code)
+            print
     elif nodetype == 'control':
         print "== Contrail Control =="
         check_status('supervisor-control', options)
