+++ src/nodemgr/config_nodemgr/contrail-config-nodemgr.conf
@@ -2,5 +2,10 @@
 server=127.0.0.1
 port=5998
 
+[DEFAULTS]
+#hostip=127.0.0.1
+#contrail_databases=config analytics
+#minimum_diskGB=4
+
 [COLLECTOR]
 #server_list=ip1:port1 ip2:port2
