+++ src/nodemgr/database_nodemgr/database_event_manager.py
@@ -7,48 +7,35 @@ monkey.patch_all()
 import os
 import sys
 import socket
-import subprocess
 import json
 import time
 import datetime
-import platform
 import select
 import gevent
-import ConfigParser
-import yaml
 
 from nodemgr.common.event_manager import EventManager
-from nodemgr.database_nodemgr.common import CassandraManager
+from nodemgr.common.cassandra_manager import CassandraManager
 
-from ConfigParser import NoOptionError
 
 from supervisor import childutils
 
 from pysandesh.sandesh_base import *
-from pysandesh.gen_py.sandesh.ttypes import SandeshLevel
-from pysandesh.sandesh_logger import SandeshLogger
 from pysandesh.sandesh_session import SandeshWriter
 from pysandesh.gen_py.sandesh_trace.ttypes import SandeshTraceRequest
 from sandesh_common.vns.ttypes import Module, NodeType
 from sandesh_common.vns.constants import ModuleNames, NodeTypeNames,\
-    Module2NodeType, INSTANCE_ID_DEFAULT, SERVICE_CONTRAIL_DATABASE, \
-    RepairNeededKeyspaces, ThreadPoolNames, UVENodeTypeNames
-from subprocess import Popen, PIPE
-from StringIO import StringIO
+    Module2NodeType, SERVICE_CONTRAIL_DATABASE, UVENodeTypeNames
 
 from nodemgr.common.sandesh.nodeinfo.ttypes import *
 from nodemgr.common.sandesh.nodeinfo.cpuinfo.ttypes import *
 from nodemgr.common.sandesh.nodeinfo.process_info.ttypes import *
 from nodemgr.common.sandesh.nodeinfo.process_info.constants import *
-from database.sandesh.database.ttypes import \
-    DatabaseUsageStats, DatabaseUsageInfo, DatabaseUsage, CassandraStatusUVE,\
-    CassandraStatusData,CassandraThreadPoolStats, CassandraCompactionTask
 from pysandesh.connection_info import ConnectionState
 
 class DatabaseEventManager(EventManager):
     def __init__(self, rule_file, discovery_server,
                  discovery_port, collector_addr,
-                 hostip, minimum_diskgb, contrail_databases,
+                 hostip, db_port, minimum_diskgb, contrail_databases,
                  cassandra_repair_interval,
                  cassandra_repair_logdir, **dss_kwargs):
         self.node_type = "contrail-database"
@@ -57,11 +44,15 @@ class DatabaseEventManager(EventManager):
         self.module = Module.DATABASE_NODE_MGR
         self.module_id = ModuleNames[self.module]
         self.hostip = hostip
+        self.db_port = db_port
         self.minimum_diskgb = minimum_diskgb
         self.contrail_databases = contrail_databases
         self.cassandra_repair_interval = cassandra_repair_interval
         self.cassandra_repair_logdir = cassandra_repair_logdir
-        self.cassandra_mgr = CassandraManager(cassandra_repair_logdir)
+        self.cassandra_mgr = CassandraManager(self.cassandra_repair_logdir,
+                                              'analyticsDb', self.contrail_databases,
+                                              self.hostip, self.minimum_diskgb,
+                                              self.db_port)
         if os.path.exists('/tmp/supervisord_database.sock'):
             self.supervisor_serverurl = "unix:///tmp/supervisord_database.sock"
         else:
@@ -95,72 +86,9 @@ class DatabaseEventManager(EventManager):
         self.third_party_process_dict["zookeeper"] = "org.apache.zookeeper.server.quorum.QuorumPeerMain"
     # end __init__
 
-    def _get_cassandra_config_option(self, config):
-        (linux_dist, x, y) = platform.linux_distribution()
-        if (linux_dist == 'Ubuntu'):
-            yamlstream = open("/etc/cassandra/cassandra.yaml", 'r')
-        else:
-            yamlstream = open("/etc/cassandra/conf/cassandra.yaml", 'r')
-
-        cfg = yaml.safe_load(yamlstream)
-        yamlstream.close()
-        return cfg[config]
-
     def msg_log(self, msg, level):
         self.sandesh_global.logger().log(SandeshLogger.get_py_logger_level(
                             level), msg)
-
-    @staticmethod
-    def cassandra_old():
-        (PLATFORM, VERSION, EXTRA) = platform.linux_distribution()
-        if PLATFORM.lower() == 'ubuntu':
-            if VERSION.find('12.') == 0:
-                return True
-        if PLATFORM.lower() == 'centos':
-            if VERSION.find('6.') == 0:
-                return True
-        return False
-
-    def process(self):
-        try:
-            cassandra_data_dirs = self._get_cassandra_config_option("data_file_directories")
-            cassandra_data_dir_exists = False
-            total_disk_space_used = 0
-            total_disk_space_available = 0
-            for cassandra_data_dir in cassandra_data_dirs:
-                if DatabaseEventManager.cassandra_old():
-                    analytics_dir = cassandra_data_dir + '/ContrailAnalytics'
-                else:
-                    analytics_dir = cassandra_data_dir + '/ContrailAnalyticsCql'
-
-                if os.path.exists(analytics_dir):
-                    cassandra_data_dir_exists = True
-                    msg = "analytics_dir is " + analytics_dir
-                    self.msg_log(msg, level=SandeshLevel.SYS_DEBUG)
-                    df = subprocess.Popen(["df", analytics_dir],
-                            stdout=subprocess.PIPE)
-                    output = df.communicate()[0]
-                    device, size, disk_space_used, disk_space_available, \
-                       percent, mountpoint = output.split("\n")[1].split()
-                    total_disk_space_used += int(disk_space_used)
-                    total_disk_space_available += int(disk_space_available)
-            if cassandra_data_dir_exists == False:
-                if 'analytics' not in self.contrail_databases:
-                    self.fail_status_bits &= ~self.FAIL_STATUS_DISK_SPACE_NA
-                else:
-                    self.fail_status_bits |= self.FAIL_STATUS_DISK_SPACE_NA
-            else:
-                disk_space_analytics = int(total_disk_space_used) + int(total_disk_space_available)
-                if (disk_space_analytics / (1024 * 1024) < self.minimum_diskgb):
-                    cmd_str = "service " + SERVICE_CONTRAIL_DATABASE + " stop"
-                    (ret_value, error_value) = Popen(
-                        cmd_str, shell=True, stdout=PIPE).communicate()
-                    self.fail_status_bits |= self.FAIL_STATUS_DISK_SPACE
-                self.fail_status_bits &= ~self.FAIL_STATUS_DISK_SPACE_NA
-        except:
-            msg = "Failed to get database usage"
-            self.msg_log(msg, level=SandeshLevel.SYS_ERR)
-            self.fail_status_bits |= self.FAIL_STATUS_DISK_SPACE_NA
 
     def send_process_state_db(self, group_names):
         self.send_process_state_db_base(
@@ -190,158 +118,6 @@ class DatabaseEventManager(EventManager):
             description += "Disk space for analytics db not retrievable."
         return description
 
-    def database_periodic(self):
-        try:
-            cassandra_data_dirs = self._get_cassandra_config_option("data_file_directories")
-            cassandra_data_dir_exists = False
-            total_disk_space_used = 0
-            total_disk_space_available = 0
-            total_analytics_db_size = 0
-            for cassandra_data_dir in cassandra_data_dirs:
-                if DatabaseEventManager.cassandra_old():
-                    analytics_dir = cassandra_data_dir + '/ContrailAnalytics'
-                else:
-                    analytics_dir = cassandra_data_dir + '/ContrailAnalyticsCql'
-
-                if os.path.exists(analytics_dir):
-                    cassandra_data_dir_exists = True
-                    msg = "analytics_dir is " + analytics_dir
-                    self.msg_log(msg, level=SandeshLevel.SYS_DEBUG)
-                    df = subprocess.Popen(["df", analytics_dir],
-                            stdout=subprocess.PIPE)
-                    output = df.communicate()[0]
-                    device, size, disk_space_used, disk_space_available, \
-                        percent, mountpoint =  output.split("\n")[1].split()
-                    total_disk_space_used += int(disk_space_used)
-                    total_disk_space_available += int(disk_space_available)
-                    du = subprocess.Popen(["du", "-skl", analytics_dir],
-                            stdout=subprocess.PIPE)
-                    analytics_db_size, directory = du.communicate()[0].split()
-                    total_analytics_db_size += int(analytics_db_size)
-            if cassandra_data_dir_exists == False:
-                if 'analytics' not in self.contrail_databases:
-                    self.fail_status_bits &= ~self.FAIL_STATUS_DISK_SPACE_NA
-                else:
-                    self.fail_status_bits |= self.FAIL_STATUS_DISK_SPACE_NA
-            else:
-                self.fail_status_bits &= ~self.FAIL_STATUS_DISK_SPACE_NA
-
-                db_stat = DatabaseUsageStats()
-                db_info = DatabaseUsageInfo()
-
-                db_stat.disk_space_used_1k = int(total_disk_space_used)
-                db_stat.disk_space_available_1k = int(total_disk_space_available)
-                db_stat.analytics_db_size_1k = int(total_analytics_db_size)
-
-                db_info.name = socket.gethostname()
-                db_info.database_usage = [db_stat]
-                usage_stat = DatabaseUsage(data=db_info)
-                usage_stat.send()
-        except:
-            msg = "Failed to get database usage"
-            self.msg_log(msg, level=SandeshLevel.SYS_ERR)
-            self.fail_status_bits |= self.FAIL_STATUS_DISK_SPACE_NA
-
-        cqlsh_cmd = "cqlsh " + self.hostip + " -e quit"
-        proc = Popen(cqlsh_cmd, shell=True, stdout=PIPE, stderr=PIPE)
-        (output, errout) = proc.communicate()
-        if proc.returncode != 0:
-            self.fail_status_bits |= self.FAIL_STATUS_SERVER_PORT
-        else:
-            self.fail_status_bits &= ~self.FAIL_STATUS_SERVER_PORT
-        self.send_nodemgr_process_status()
-        # Send cassandra nodetool information
-        self.send_database_status()
-        # Record cluster status and shut down cassandra if needed
-        self.cassandra_mgr.status()
-    # end database_periodic
-
-    def send_database_status(self):
-        cassandra_status_uve = CassandraStatusUVE()
-        cassandra_status = CassandraStatusData()
-        cassandra_status.cassandra_compaction_task = CassandraCompactionTask()
-        # Get compactionstats
-        compaction_count = subprocess.Popen("nodetool compactionstats|grep 'pending tasks:'",
-            shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        op, err = compaction_count.communicate()
-        if compaction_count.returncode != 0:
-            msg = "Failed to get nodetool compactionstats " + err
-            self.msg_log(msg, level=SandeshLevel.SYS_ERR)
-            return
-        cassandra_status.cassandra_compaction_task.pending_compaction_tasks = \
-            self.get_pending_compaction_count(op)
-        # Get the tpstats value
-        tpstats_op = subprocess.Popen(["nodetool", "tpstats"], stdout=subprocess.PIPE,
-                                      stderr=subprocess.PIPE)
-        op, err = tpstats_op.communicate()
-        if tpstats_op.returncode != 0:
-            msg = "Failed to get nodetool tpstats " + err
-            self.msg_log(msg, level=SandeshLevel.SYS_ERR)
-            return
-        cassandra_status.thread_pool_stats = self.get_tp_status(op)
-        cassandra_status.name = socket.gethostname()
-        cassandra_status_uve = CassandraStatusUVE(data=cassandra_status)
-        if self.has_cassandra_status_changed(cassandra_status, self.cassandra_status_old):
-            # Assign cassandra_status to cassandra_status_old
-            self.cassandra_status_old.thread_pool_stats = \
-                cassandra_status.thread_pool_stats
-            self.cassandra_status_old.cassandra_compaction_task.\
-                pending_compaction_tasks = cassandra_status.\
-                cassandra_compaction_task.pending_compaction_tasks
-            msg = 'Sending UVE: ' + str(cassandra_status_uve)
-            self.sandesh_global.logger().log(SandeshLogger.get_py_logger_level(
-                            SandeshLevel.SYS_DEBUG), msg)
-            cassandra_status_uve.send()
-    # end send_database_status
-
-    def has_cassandra_status_changed(self,current_status, old_status):
-        if current_status.cassandra_compaction_task.pending_compaction_tasks != \
-            old_status.cassandra_compaction_task.pending_compaction_tasks :
-            return True
-        i = 0
-        if len(current_status.thread_pool_stats) != \
-            len(old_status.thread_pool_stats):
-            return True
-        while i < len(current_status.thread_pool_stats):
-            if (current_status.thread_pool_stats[i].active != \
-                old_status.thread_pool_stats[i].active or
-                current_status.thread_pool_stats[i].pending != \
-                old_status.thread_pool_stats[i].pending or
-                current_status.thread_pool_stats[i].all_time_blocked != \
-                old_status.thread_pool_stats[i].all_time_blocked):
-                return True
-            i = i+1
-        return False
-    # end has_cassandra_status_changed
-
-    def get_pending_compaction_count(self, pending_count):
-        compaction_count_val = pending_count.strip()
-        # output is of the format pending tasks: x
-        pending_count_val = compaction_count_val.split(':')
-        return int(pending_count_val[1].strip())
-    # end get_pending_compaction_count
-
-    def get_tp_status(self,tp_stats_output):
-        tpstats_rows = tp_stats_output.split('\n')
-        thread_pool_stats_list = []
-        for row_index in range(1, len(tpstats_rows)):
-            cols = tpstats_rows[row_index].split()
-            # If tpstats len(cols) > 2, else we have reached the end
-            if len(cols) > 2:
-                if (cols[0] in ThreadPoolNames):
-                    # Create a CassandraThreadPoolStats for matching entries
-                    tpstat = CassandraThreadPoolStats()
-                    tpstat.pool_name = cols[0]
-                    tpstat.active = int(cols[1])
-                    tpstat.pending = int(cols[2])
-                    tpstat.all_time_blocked = int(cols[5])
-                    thread_pool_stats_list.append(tpstat)
-            else:
-                # Reached end of tpstats, breaking because dropstats follows
-                break
-        return thread_pool_stats_list
-    # end get_tp_status
-
     def runforever(self, test=False):
         self.prev_current_time = int(time.time())
         # Initialize tpstat structures
@@ -369,9 +145,13 @@ class DatabaseEventManager(EventManager):
                 self.event_process_communication(pdata)
             # do periodic events
             if headers['eventname'].startswith("TICK_60"):
-                self.database_periodic()
+                self.cassandra_mgr.database_periodic(self)
                 self.event_tick_60()
             self.listener_nodemgr.ok(self.stdout)
 
     def get_package_name(self):
         return self.node_type + '-common'
+
+    def process(self):
+        self.cassandra_mgr.process(self)
+    # end process
